refactor(stack): split deployer.go into focused domain files
Move lifecycle operations (create/update/delete/poll) to deployer_lifecycle.go
and helpers (constants, describeStack, readFleetIDFromOutputs, isStackNotFound,
stackDeletionPollResult) to deployer_helpers.go. deployer.go retains only the
public API (Deploy, Status, Destroy, GetFleetID, CreateGameSession,
DescribeGameSession) and type definitions. No behavior changes.
